+++ src/objects/FleetShip.cpp
@@ -8,7 +8,7 @@
 		ShipData::ShipData *data = DataHandler.getShipById(this->getTypeId());
 		double structure = data->getStructure() + data->getShield();
 		if (!structure) this->count = 0;
-		int shipCnt = ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
+		int shipCnt = (int)ceil((this->initCount - this->count) * config.nget("ship_wf_percent",0));
 		
 		this->rebuildIsCalced = true;
 		
@@ -22,7 +22,7 @@
 		ShipData::ShipData *data = DataHandler.getShipById(this->getTypeId());
 		double structure = data->getStructure() + data->getShield();
 		if (!structure) this->count = 0;
-		int shipCnt = ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
+		int shipCnt = (int)ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
 		
 		this->rebuildIsCalced = true;
 		
@@ -36,7 +36,7 @@
 		ShipData::ShipData *data = DataHandler.getShipById(this->getTypeId());
 		double structure = data->getStructure() + data->getShield();
 		if (!structure) this->count = 0;
-		int shipCnt = ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
+		int shipCnt = (int)ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
 		
 		this->rebuildIsCalced = true;
