+++ src/objects/ListShip.cpp
@@ -7,7 +7,7 @@
 		
 		DataHandler &DataHandler = DataHandler::instance();
 		ShipData::ShipData *data = DataHandler.getShipById(this->getTypeId());
-		int shipCnt = ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
+		int shipCnt = (int)ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
 		
 		this->rebuildIsCalced = true;
 		
@@ -19,7 +19,7 @@
 		
 		DataHandler &DataHandler = DataHandler::instance();
 		ShipData::ShipData *data = DataHandler.getShipById(this->getTypeId());
-		int shipCnt = ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
+		int shipCnt = (int)ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
 		
 		this->rebuildIsCalced = true;
 		
@@ -31,7 +31,7 @@
 		
 		DataHandler &DataHandler = DataHandler::instance();
 		ShipData::ShipData *data = DataHandler.getShipById(this->getTypeId());
-		int shipCnt = ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
+		int shipCnt = (int)ceil((this->initCount - this->count)*config.nget("ship_wf_percent",0));
 		
 		this->rebuildIsCalced = true;
